refactor(keyframes): split dopesheet and graph-interaction into sibling modules
Extract self-contained pieces out of the two largest files in the
keyframes feature with zero behavior change. Pulls icon and zoom-control
sub-components, layout constants, dopesheet types, grouping/tick
helpers, localStorage persistence for visible graph properties, and the
graph-interaction types/constants into focused sibling files so the
remaining component and hook bodies are easier to navigate.
